breast-cancer-rag-app/backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
flask-limiter==2.8.1
openai==0.28.1
notebookutils==1.0.0
pyspark==3.4.1
azure-monitor-opentelemetry==1.0.0
tenacity==8.2.2
gunicorn==20.1.0